fix: AI not replying in debates
This commit fixes an issue where the AI was not replying in debates.

- The `user_message` and `end_debate` event handlers were moved from `debate.py` to `matchmaking.py`.
- The `get_ai_response` and `evaluate_debate` functions are now correctly imported in `matchmaking.py`.
