+++ backend/app/ai.py
@@ -4,6 +4,7 @@ import os
 GROQ_API_KEY = os.getenv("GROQ_API_KEY")
 
 def get_ai_response(prompt: str) -> str:
+    print(f"Getting AI response for prompt: {prompt}")
     headers = {
         "Authorization": f"Bearer {GROQ_API_KEY}",
         "Content-Type": "application/json"
@@ -20,6 +21,9 @@ def get_ai_response(prompt: str) -> str:
     response = requests.post("https://api.groq.com/openai/v1/chat/completions", headers=headers, json=body)
 
     if response.status_code == 200:
+        print("AI response received successfully.")
         return response.json()['choices'][0]['message']['content']
     else:
+        print(f"AI failed to respond with status code: {response.status_code}")
+        print(f"Response: {response.text}")
         return "AI failed to respond."
